style: Extract from RestaurantCard components
