+++ src/components/promotion/PromotionDetails.less
@@ -44,7 +44,7 @@
   }
 
   .promotion-description {
-    -webkit-line-clamp: 3;
+    -webkit-line-clamp: 1;
     -webkit-box-orient: vertical;
     display: -webkit-box;
     font-size: 0.8rem;
