+++ src/components/restaurant/RestaurantCard.less
@@ -23,6 +23,12 @@
         margin-top: 0px;
       }
 
+      .close-card {
+        font-size: 1.5em;
+        color: @gray;
+        float: right;
+      }
+
       .restaurant-card-title {
         font-family: 'AbeeZee';
         font-weight: bold;
