+++ src/components/restaurant/card/Header.less
@@ -1,0 +1,37 @@
+@import '../../../variables.less';
+
+.restaurant-details {
+  align-items: center;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+
+  .action-button {
+    border: 1px solid @yellow;
+    border-radius: 20px;
+    box-shadow: 0 2px 4px 0 fade(@black, 20%);
+    font-size: 0.9em;
+    font-weight: bold;
+    margin-right: 10px;
+    padding-left: 8px;
+    padding-right: 8px;
+
+    &:hover {
+      color: @black;
+    }
+  }
+
+  .rating {
+    font-size: 1.5em;
+    margin-bottom: 5px;
+    padding: 0px;
+  }
+
+  .restaurant-details-text {
+    color: @gray;
+    display: inline;
+    font-weight: bold;
+    margin: 10px 10px 0px 0px;
+    font-size: 0.8rem;
+  }
+}
