+++ src/components/restaurant/card/Header.tsx
@@ -1,10 +1,9 @@
-import './Header.css';
+import './Header.less';
 
 import { Button, Rate, Row } from 'antd';
 import React, { ReactElement } from 'react';
 
 interface Props {
-  // cuisine: string; // todo: this should use a promotions cuisine https://promopal.atlassian.net/browse/PP-91
   distance: number;
   name: string | undefined;
   priceLevel: number | undefined;
@@ -43,39 +42,26 @@ export default function Header(props: Props): ReactElement {
 
   const separator = <p className="restaurant-details-text">&#8226;</p>;
 
-  const Buttons = (): ReactElement => (
-    <Row className="action-buttons restaurant-card-header-component">
-      <Button className="action-button">
-        <a href={props.website} target="_blank" rel="noopener noreferrer">
-          Website
-        </a>
-      </Button>
-      {/* <Button className="action-button">
-        <a href="/">Directions</a>
-      </Button>
-      <Button className="action-button">
-        <a href="/">Call</a>
-      </Button> */}
-    </Row>
-  );
-
   return (
-    <>
-      <Row className="restaurant-card-header-component">
+    <Row className="restaurant-card-header-component restaurant-details">
+      <div>
         <Rate
           allowHalf={true}
           disabled={true}
           className="rating"
           defaultValue={normalizeRating(props.rating)}
         />
-      </Row>
-      <Row className="restaurant-card-header-component">
-        {/*<p className="restaurant-details-text">{props.cuisine}</p>*/}
-        <p className="restaurant-details-text">{formatPriceLevel(props.priceLevel)}</p>
-        {separator}
-        <p className="restaurant-details-text">{props.distance}m</p>
-      </Row>
-      <Buttons />
-    </>
+        <div>
+          <p className="restaurant-details-text">{formatPriceLevel(props.priceLevel)}</p>
+          {separator}
+          <p className="restaurant-details-text">{props.distance}m</p>
+        </div>
+      </div>
+      <Button className="action-button">
+        <a href={props.website} target="_blank" rel="noopener noreferrer">
+          Website
+        </a>
+      </Button>
+    </Row>
   );
 }
