--- src/components/restaurant/card/tab/PromotionsTab.css
@@ -1,95 +1,0 @@
-/* Clock icon */
-.promotion-info .clock-icon {
-  color: #8b8888;
-  font-size: 0.9em;
-}
-
-/* Promotion description */
-.promotion-info .description {
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  display: -webkit-box;
-  margin-top: 5px;
-  margin-bottom: 5px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  word-break: break-word;
-}
-
-/* Promotion expiration date */
-.promotion-info .expiration-date p {
-  font-weight: bold;
-}
-
-/* Time */
-.promotion-info .schedules-time {
-  margin: 0;
-}
-
-/* Promotions tab */
-.tab-promotions {
-  margin-bottom: 10px;
-}
-
-/* Promotions tab paragraph text */
-.tab-promotions p {
-  margin-bottom: 0;
-}
-
-/* Promotions schedules text */
-.tab-promotions .schedules {
-  color: #8b8888;
-  font-size: 0.8rem;
-}
-
-/* Promotions empty schedules text */
-.tab-promotions .schedules-empty {
-  text-align: center;
-}
-
-/* Promotions tab content */
-.tab-promotions .ant-tabs-content.ant-tabs-content-top {
-  margin-bottom: 0;
-}
-
-/* Promotions schedules */
-.tab-promotions .ant-tabs-content-holder {
-  max-height: 200px;
-  overflow: scroll;
-  padding-right: 15px;
-}
-
-/* Tab bar */
-.tab-promotions .ant-tabs-nav-wrap {
-  background-color: #fff;
-  margin-top: 0px;
-  margin-bottom: -10px;
-}
-
-/* Underline on active tab */
-.tab-promotions .ant-tabs-nav-list .ant-tabs-ink-bar.ant-tabs-ink-bar-animated {
-  height: 2px;
-  background: #fac15e;
-}
-
-/* Active tab bar */
-.tab-promotions .ant-tabs-nav-list .ant-tabs-tab.ant-tabs-tab-active {
-  background-color: #fff;
-}
-
-/* Active tab bar button text */
-.tab-promotions .ant-tabs-nav-list .ant-tabs-tab.ant-tabs-tab-active .ant-tabs-tab-btn {
-  font-weight: 600;
-}
-
-/* Tab bar button spacing */
-.tab-promotions .ant-tabs-nav-wrap .ant-tabs-nav-list .ant-tabs-tab {
-  margin-right: 0px;
-  padding-left: 10px;
-  padding-right: 10px;
-}
-
-/* Tab bar button text */
-.tab-promotions .ant-tabs-nav-list .ant-tabs-tab .ant-tabs-tab-btn {
-  color: #8b8888;
-}
