+++ src/variables.less
@@ -3,6 +3,7 @@
 @light-gray: #f5f5f5;
 @orange: #ffa200;
 @red: #f7263f;
+@white: #ffffff;
 @yellow: #ffc529;
 
 @shadow: 0px 2px 10px fade(@black, 20%);
